+++ RELEASE_NOTES.md
@@ -1,18 +1,20 @@
 # Release Notes
 
 ## 0.52
-- Moving to use DNAnexus workflow inputs and outputs.
-
+- Porting the code to use DNAnexus workflow inputs and outputs.
 Running a workflow now has slighly easier command line syntax.
 For example, if workflow `foo` takes an integer
 argument `N`, then running it through the CLI is done like
 this:
+
 ```
 dx run foo -iN=19
 ```
 
 - Revamp the conversions between dx:file and wdl:file. This allows
 specifying dx files in defaults.
+- Initial support for non-empty WDL arrays (Array[Int]+)
+
 
 ## 0.51
 - Adding a `-quiet` flag to suppress warning and informational